XMLHttpRequest/tests/submissions/Opera/open-after-abort.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: open() after abort()</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest(),
            control_flag = false,
            result = [],
            expected = [1, 1, 4, 1] // open() -> 1, send() -> 1,
                                    // abort() -> 4, open() -> 1
        client.onreadystatechange = function() {
          test.step(function() {
            result.push(client.readyState)
          })
        }
        client.open("GET", "resources/delay.php?ms=2000", true)
        client.send(null)
        client.abort()
        assert_equals(client.readyState, 0)
        client.open("GET", "resources/delay.php?ms=2000", true)
        assert_array_equals(result, expected)
      })
      test.done()
    </script>
  </body>
</html>
